// src/interfaces/item.ts
export type Item = {
  id: number;
  name: string;
  price: string;
  brand: string;
  rating: number;
  category: null | string;
  currency: null | string;
  description: string;
  image_link: string;
  price_sign: null | string;
  product_api_url: string;
  product_colors: any[];
  product_link: string;
  product_type: string;
  tag_list: any[];
  website_link: string;
  api_featured_image: string;
  created_at: string;
  updated_at: string;
};